Numpad 9 now resets speedOffset; Fixed (?) some collision issues such as getting stuck into walls. Water now has different levels that are encoded in metadata. Tiles now need to take damage before breaking instead of breaking instantly, damage is also encoded in meta. Game now runs in fullscreen mode by default. Fixed bug that crashed the game when walking on tilled sand.

+++ Entities/MovingEntity.cpp
@@ -3,7 +3,7 @@
 
 void MovingEntity::Update(World* world, float dt)
 {
-	inWater = (world->TileAt(position.x + .5f, position.y + .5f, 0) == TileID::Water);
+	inWater = (world->TileAt((int) (position.x + .5f), (int) (position.y + .5f), 0) == TileID::Water);
 
 	float actualSpeed = inWater ? speed / 2 : speed;
